+++ src/css-rules/no-unitless-values/index.ts
@@ -1,0 +1,3 @@
+import noUnitlessValuesRule from './rule-definition.js';
+
+export default noUnitlessValuesRule;